fix imports after renaming
