+++ components/admin/data/layers/form/interactions/reducer.js
@@ -1,4 +1,4 @@
-import * as actions from './interactions-actions';
+import * as actions from './actions';
 
 export const initialState = {
   added: [],
